API

This page contains documentation of the public API of MRIReco. In the Julia REPL one can access this documentation by entering the help mode with ? and then writing the function for which the documentation should be shown.

Operators

Operators are implemented as subtypes of AbstractLinearOperator, which is defined in the package LinearOperators.jl. Such operators must provide a function implementing the product and a function implementing the product with the adjoint. Furthermore, the number of rows and columns of the operator must be specified.

Missing docstring.

Missing docstring for MRIReco.encodingOps2d_simple. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOps3d_simple. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOps2d_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOps3d_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOp2d_multiEcho. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOp3d_multiEcho. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOp2d_multiEcho_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.encodingOp3d_multiEcho_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.fourierEncodingOp2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.fourierEncodingOp3d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.ExplicitOp(shape::NTuple{D,Int64}, tr::Trajectory, correctionmap::Array{ComplexF64,D}; MRIReco.echoImage::Bool=false, kargs...) where D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.FFTOp(T::Type, shape::Tuple, shift=true). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.NFFTOp(shape::Tuple, tr::Trajectory; nodes=nothing, kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.FieldmapNFFTOp(shape::NTuple{D,Int64}, tr::Trajectory, correctionmap::Array{ComplexF64,D}; method::String="nfft", echoImage::Bool=true, alpha::Float64=1.75, m::Float64=3.0, K=20, kargs...) where D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.SamplingOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.SensitivityOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.SparseOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.RegularizedLeastSquares.WeightingOp. Check Documenter's build log for details.

Datatypes

Missing docstring.

Missing docstring for MRIReco.AcquisitionData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.AcquisitionData(tr::T,kdata::Array{Matrix{ComplexF64},3} ; seqInfo=Dict{Symbol,Any}() , idx=nothing , encodingSize=Int64[0,0,0] , fov=Float64[0,0,0] , kargs...) where T <: Union{Trajectory,Vector{Trajectory}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.trajectory(acqData::AcquisitionData,i::Int64=1). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numContrasts(acqData::AcquisitionData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numChannels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numSlices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numRepetitions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.kData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.multiEchoData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.multiCoilData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.multiCoilMultiEchoData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.profileData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.samplingDensity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.changeEncodingSize2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.convert3dTo2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.RawAcquisitionData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.trajectory(f::RawAcquisitionData; slice::Int=1, contrast::Int=1). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.rawdata(f::RawAcquisitionData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.AcquisitionData(f::RawAcquisitionData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.RawAcquisitionData(f::ISMRMRDFile, dataset="dataset"). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.AcquisitionData(f::ISMRMRDFile, dataset="dataset"). Check Documenter's build log for details.

Trajectories

Missing docstring.

Missing docstring for MRIReco.Trajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.trajectory(trajName::AbstractString, numProfiles::Int, numSamplingPerProfile::Int; MRIReco.numSlices::Int64=1, TE::Float64=0.0, AQ::Float64=1.e-3, kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.string(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.echoTime(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.acqTimePerProfile(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numProfiles(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numSamplingPerProfile(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numSlices(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.isCircular(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.isCartesian(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.dims(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.kspaceNodes(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.readoutTimes(tr::Trajectory). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.CartesianTrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.EPITrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.OneLine2dTrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.RadialTrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.SpiralTrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.SpiralTrajectoryVarDens. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.CartesianTrajectory3D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.KooshballTrajectory. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.StackOfStarsTrajectory. Check Documenter's build log for details.

Sequences

Missing docstring.

Missing docstring for MRIReco.MESequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.numContrasts(seq::MESequence). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.echoTimes(seq::MESequence). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.flipAngles(seq::MESequence). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.echoAmplitudes(seq::MESequence, R1::Float64, R2::Float64, numStates=nothing). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.epgAmplitudes(seq::MESequence, R1::Real, R2::Real, numStates=nothing). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.epgRotation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.epgRelaxation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.epgDephasing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.rfRotation. Check Documenter's build log for details.

Sampling

Missing docstring.

Missing docstring for MRIReco.sample. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_kspace(data::AbstractArray,redFac::Float64,patFunc::AbstractString;kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_kspace(acqData::AcquisitionData,redFac::Float64, patFunc::AbstractString; rand=true, profiles=true, seed = 1234, kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_regular(shape::Tuple, redFac::Float64; kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_random(shape::Tuple{Int64,Int64},redFac::Float64;calsize::Int64=0,kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_poissondisk(shape::Tuple{Int64,Int64},redFac::Float64;calsize::Int64=0, seed::Int64=1234,kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_vdpoisson(shape::Tuple{Int64,Int64},redFac::Float64; seed::Int64=1234,kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.sample_lines(shape::Tuple{Int64,Int64},redFac::Float64;sampleFunc="random",kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.calculateIncoherence(acqData::AcquisitionData, recoParams::Dict, slice=1). Check Documenter's build log for details.

Simulation

Missing docstring.

Missing docstring for MRIReco.simulation(image::Array{T,3}, simParams::Dict) where T<:Union{ComplexF64,Float64}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.simulation(image::Array{T,3}, simParams::Dict, filename::String; force=false) where T<:Union{ComplexF64,Float64}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.simulation(image::Array{T,2}, simParams::Dict) where T<:Union{ComplexF64,Float64}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.simulation(tr::Trajectory , image::Array{ComplexF64} , correctionMap = [] ; opName="fast" , senseMaps=[] , verbose=true , kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.simulation(seq::AbstractSequence, tr::Vector{Trajectory} , image::Array{ComplexF64,3} ; opName="fast" , r1map=[] , r2map=[] , fmap=[] , senseMaps=[] , verbose=true , kargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.addNoise(x::Vector, snr::Float64, complex= true). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.addNoise(acqData::AcquisitionData, snr::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.addNoise!(acqData::AcquisitionData, snr::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.birdcageSensitivity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.measured2DSensitivity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.quadraticFieldmap. Check Documenter's build log for details.

Reconstruction

MRIReco.reconstruction — Method

reconstruction(acqData::AcquisitionData, recoParams::Dict)

Performs image reconstruction of an AcquisitionData object. Parameters are specified in a dictionary.

Reconstruction types are specified by the symbol :reco. Valid reconstruction names are:

:direct - direct Fourier reconstruction

:standard - iterative reconstruction for all contrasts, coils & slices independently

:multiEcho - iterative joint reconstruction of all echo images

:multiCoil - SENSE-type iterative reconstruction

:multiCoilMultiEcho - SENSE-type iterative reconstruction of all echo images

MRIReco.reconstruction — Method

reconstruction(acqData::AcquisitionData, recoParams::Dict,filename::String; force=false)

performs the same image reconstrucion as reconstruction(acqData::AcquisitionData, recoParams::Dict) and saves the image in a file with name filename. If force=false, the reconstructed image is loaded from the the file filename if the latter is present.

Missing docstring.

Missing docstring for MRIReco.setupIterativeReco. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.reconstruction_direct_2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRIReco.reconstruction_direct_3d. Check Documenter's build log for details.

MRIReco.reconstruction_simple — Function

Performs iterative image reconstruction independently for the data of all coils, contrasts and slices

Arguments

acqData::AcquisitionData - AcquisitionData object

reconSize::NTuple{2,Int64} - size of image to reconstruct

reg::Vector{<:AbstractRegularization} - Regularization to be used

sparseTrafo::AbstractLinearOperator - sparsifying transformation

weights::Vector{Vector{Complex{<:AbstractFloat}}} - sampling density of the trajectories in acqData

solver::Type{<:AbstractLinearSolver} - name of the solver to use

(normalize::Bool=false) - adjust regularization parameter according to the size of k-space data

(params::Dict{Symbol,Any}) - Dict with additional parameters

MRIReco.reconstruction_multiEcho — Function

Performs a iterative image reconstruction jointly for all contrasts. Different slices and coil images are reconstructed independently.

Arguments

acqData::AcquisitionData - AcquisitionData object

reconSize::NTuple{2,Int64} - size of image to reconstruct

reg::Vector{<:AbstractRegularization} - Regularization to be used

sparseTrafo::AbstractLinearOperator - sparsifying transformation

weights::Vector{Vector{Complex{<:AbstractFloat}}} - sampling density of the trajectories in acqData

solver::Type{<:AbstractLinearSolver} - name of the solver to use

(normalize::Bool=false) - adjust regularization parameter according to the size of k-space data

(params::Dict{Symbol,Any}) - Dict with additional parameters

MRIReco.reconstruction_multiCoil — Function

Performs a SENSE-type iterative image reconstruction. Different slices and contrasts images are reconstructed independently.

Arguments

acqData::AcquisitionData - AcquisitionData object

reconSize::NTuple{2,Int64} - size of image to reconstruct

reg::Vector{<:AbstractRegularization} - Regularization to be used

sparseTrafo::AbstractLinearOperator - sparsifying transformation

weights::Vector{Vector{Complex{<:AbstractFloat}}} - sampling density of the trajectories in acqData

L_inv::Array{Complex{<:AbstractFloat}} - noise decorrelation matrix

solver::Type{<:AbstractLinearSolver} - name of the solver to use

senseMaps::AbstractArray{Complex{<:AbstractFloat}} - coil sensitivities

(normalize::Bool=false) - adjust regularization parameter according to the size of k-space data

(params::Dict{Symbol,Any}) - Dict with additional parameters

MRIReco.reconstruction_multiCoilMultiEcho — Function

Performs a SENSE-type iterative image reconstruction which reconstructs all contrasts jointly. Different slices are reconstructed independently.

Arguments

acqData::AcquisitionData - AcquisitionData object

reconSize::NTuple{2,Int64} - size of image to reconstruct

reg::Vector{<:AbstractRegularization} - Regularization to be used

sparseTrafo::AbstractLinearOperator - sparsifying transformation

weights::Vector{Vector{Complex{<:AbstractFloat}}} - sampling density of the trajectories in acqData

solver::Type{<:AbstractLinearSolver} - name of the solver to use

senseMaps::AbstractArray{Complex{<:AbstractFloat}} - coil sensitivities

(normalize::Bool=false) - adjust regularization parameter according to the size of k-space data

(params::Dict{Symbol,Any}) - Dict with additional parameters

Missing docstring.

Missing docstring for MRIReco.espirit. Check Documenter's build log for details.

MRIReco.nrmsd — Function

nrmsd(I,Ireco)

computes the normalized root mean squared error of the image Ireco with respect to the image I.